splitting lexer tests
